collector/requirements.txt
flask==3.0.0
flask-cors==4.0.0
akshare==1.11.0
requests==2.31.0
apscheduler==3.10.4
pymysql==1.1.0
sqlalchemy==2.0.23
redis==5.0.1
python-dotenv==1.0.0
loguru==0.7.2
